+++ src/app/client/widget/homepage-client/homepage-client.component.ts
@@ -4,6 +4,7 @@ import { Menu } from 'src/app/core/models/menu.model';
 import { HttpResponse } from '@angular/common/http';
 import { ClientService } from 'src/app/core/services/client.service';
 import { SessionStorageService } from 'src/app/core/services/session-storage.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-homepage-client',
@@ -14,16 +15,22 @@ export class HomepageClientComponent implements OnInit {
   restoList: RestaurantModel[] = []
   menuTest = new Menu();
 
-  constructor(private sessionStorageService: SessionStorageService,private clientService: ClientService) { };
+  constructor(private toastr: ToastrService, private sessionStorageService: SessionStorageService, private clientService: ClientService) { };
 
   ngOnInit(): void {
     let token = this.sessionStorageService.getItem('token');
-    this.clientService.getRestaurant(token).subscribe((response: RestaurantModel) => {
-      for(let i = 0;i<response.restaurants.length;i++){
-        console.log(response.restaurants[i])
-        let restaurantInformation = response.restaurants[i];
-        this.restoList.push(restaurantInformation);
+    this.clientService.getRestaurant(token).subscribe({
+      next: (response: RestaurantModel) => {
+        for (let i = 0; i < response.restaurants.length; i++) {
+          let restaurantInformation = response.restaurants[i];
+          this.restoList.push(restaurantInformation);
+        }
+      },
+      error: () => {
+        this.toastr.error("Erreur lors de la récupération du token ");
       }
     });
+
+
   }
 }