+++ src/app/client/widget/restaurant-client/restaurant-client.component.ts
@@ -1,8 +1,10 @@
-import { HttpResponse } from '@angular/common/http';
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
+import { InfoRestaurantModel } from 'src/app/core/models/infoRestaurant.model';
 import { Menu } from 'src/app/core/models/menu.model';
 import { ClientService } from 'src/app/core/services/client.service';
+import { RestaurantService } from 'src/app/core/services/restaurant.service';
 import { SessionStorageService } from 'src/app/core/services/session-storage.service';
 
 @Component({
@@ -10,17 +12,21 @@ import { SessionStorageService } from 'src/app/core/services/session-storage.ser
   templateUrl: './restaurant-client.component.html',
   styleUrls: ['./restaurant-client.component.css']
 })
-export class RestaurantClientComponent {
+export class RestaurantClientComponent  implements OnInit{
   restaurant!: string;
   address!: string;
   menuList: Menu[] = [];
-  menuTest = new Menu();
   articleList: Menu[] = [];
-  articleTest = new Menu();
   type!: string | null;
-  idRestaurant!: string;
+  idRestaurant!: string|null;
 
-  constructor(private clientService: ClientService, private sessionStorageService: SessionStorageService, private router: Router, private route: ActivatedRoute) { }
+  constructor(
+    private toastr: ToastrService,
+    private clientService: ClientService,
+    private sessionStorageService: SessionStorageService,
+    private router: Router,
+    private route: ActivatedRoute,
+  ) { }
 
   ngOnInit(): void {
     this.type = this.sessionStorageService.getItem('type');
@@ -29,28 +35,27 @@ export class RestaurantClientComponent {
     }
 
 
-    this.addMenu();
+    this.addValue();
   }
 
-  addMenu() {
+  addValue() {
     let token = this.sessionStorageService.getItem('token');
-    this.clientService.getRestaurantDetail(this.idRestaurant, token).subscribe((response: HttpResponse<any>) => {
-      console.log(response)
+    this.route.paramMap.subscribe(params => {
+      // Récupérer la valeur de resto.id
+      this.idRestaurant = params.get('id');
     });
-    this.restaurant = "Mcgronald’s";
-    this.address = "243 rue de la Republique , 95239";
 
-    this.menuTest.img = "https://eu-images.contentstack.com/v3/assets/blt5004e64d3579c43f/blt9418bc6e38e6544a/660439b26eb29729ab905cdf/MxBO_BIGMAC.png?auto=webp&width=1280&disable=upscale";
-    this.menuTest.description = "Burger maison, boisson, frites";
-    this.menuTest.name = "Menu ElClassico"
-    this.menuTest.price = "12,40€";
-    this.menuList.push(this.menuTest);
-    this.menuList.push(this.menuTest);
-
-    this.articleTest.img = "https://eu-images.contentstack.com/v3/assets/blt5004e64d3579c43f/blta639ebc798204a17/66043a77dd76c70108f663a3/400x400_Big_Mac.png?auto=webp&width=1280&disable=upscale";
-    this.articleTest.description = "Bun's, Steak, Salade, Fromage";
-    this.articleTest.name = "Big Muc"
-    this.articleTest.price = "5€40";
-    this.articleList.push(this.articleTest);
+    this.clientService.getRestaurantDetail(this.idRestaurant, token).subscribe({
+      next: (response: InfoRestaurantModel) => {   
+        let value = response.restaurant;
+        this.restaurant = value.name;
+        this.address = value.address;
+        this.articleList = value.products;
+        this.menuList = value.menus;
+      },
+      error: () => {
+        this.toastr.error("Erreur lors de la récupération du détail des restaurants ");
+      }
+    });
   }
 }