+++ src/app/core/components/card/card.component.ts
@@ -1,37 +1,39 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Input } from '@angular/core';
 import { RestaurantModel } from '../../models/restaurant.model';
 import { Menu } from '../../models/menu.model';
+import { PanierModel } from '../../models/panier-create.model';
+import { PanierService } from '../../services/panier.service';
 
 @Component({
   selector: 'app-card',
   templateUrl: './card.component.html',
   styleUrls: ['./card.component.css']
 })
-export class CardComponent {
+export class CardComponent implements OnInit {
 
-  @Input() restaurant!:RestaurantModel;
-  @Input() menu!:Menu;
-  @Input() isDetailResto!:boolean;
+
+  panier: PanierModel[] = [];
+  product!: PanierModel;
+
+  constructor(private panierService: PanierService) { }
+
+  @Input() restaurant!: RestaurantModel;
+  @Input() menu!: Menu;
+  @Input() isDetailResto!: boolean;
   @Input() isClient!: boolean;
+  @Input() isMenu!: boolean;
+  @Input() idRestaurant!: any;
 
-  quantity !: number;
+  ngOnInit(): void {
+  }
 
-  decrease = document.querySelector('.decrease');
-  increase = document.querySelector('.increase');
-  text = document.querySelector('.text-quantity');
-  num=0;
-  
-  
-  onChangePlus(){
-    console.log('+')
-    this.num +=1;
-    this.text!.innerHTML=this.num.toString();
+
+  onClick(id: any, isMenu: boolean, name: any, price: any, haveDrink: boolean) {
+    var idRestaurantValue = this.idRestaurant;
+    console.log(id);
+    this.panierService.addPanier({ id,idRestaurantValue, isMenu, name, price, haveDrink, drink: "" })
+    console.log(this.panierService.getPanier());
   }
-  
-  onChangeMinus(){
-    console.log('-')
-    this.num -=1;
-    this.text!.innerHTML=this.num.toString();
-  }
+
 }