+++ src/app/core/services/apiPaths.ts
@@ -6,7 +6,9 @@ export const API = Object.freeze({
     login: '/auth/login',
 
     logs:'/auth/logs',
-
+    getComponentLog:'/component/getLogs',
+    writeComponentLog:'/component/writeLogs',
+    
     editUser: '/user/edit',
     deleteUser: '/user/delete',
     activateUser: '/user/unsuspend',
@@ -14,18 +16,38 @@ export const API = Object.freeze({
     getUserInfo: '/user/get',
     getUsers: '/user/getAll',
 
+    createOrder: '/order/create',
+    updateOrder: '/order/update',
     getOrder: '/order/getOrder',
     getOrders: '/order/getOrders',
     getAllOrders: '/order/getAllOrders',
+    getAllOrdersFromRestaurant: '/order/getAllOrdersFromRestaurant',
+    getAllCreatedOrdersFromRestaurant: '/order/getAllCreatedOrdersFromRestaurant',
+
 
     createMenu:'/menu/create',
     createArticle:'/product/add',
+    getMenu:'/menu/find',
+    getArticle:'/product/find',
+    getDrinks:'/product/getDrinks',
 
+
+
+    getOrdersSince:'/restaurant/getOrdersSince',
     getRestaurantInformation: '/restaurant/findByOwner',
     getRestaurant:'/restaurant/getAll',
     getRestaurantDetail:'/restaurant/find',
     getOrderCount: '/restaurant/getOrderCount',
 
+
+    getDelivery:'/delivery/getAllWithFilter',
+    refuseDelivery:'/delivery/refuse',
+    acceptDelivery:'/delivery/accept',
+    getStatutDelivery:'/delivery/getStatut',
+    trackDeliveryDelivery:'/delivery/trackDelivery',
+    nerbyDelivery:'/delivery/nearby',
+    validateDelivery:'/delivery/validate',
+    getMetricDelivery:'/delivery/metrics',