+++ src/app/core/services/restaurant.service.ts
@@ -21,6 +21,16 @@ export class RestaurantService {
         return this.http.get(`${environment.urlRestaurant}${API.getRestaurantInformation}`+ '?id=' + userID, httpOptions );
     }
 
+    getOrdersNumbers(token: any,restaurantID:any,daysBack:any): Observable<any> {
+        const httpOptions = {
+            headers: new HttpHeaders({
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`
+            })
+        };
+        return this.http.get(`${environment.urlRestaurant}${API.getOrdersSince}`+ '?restaurantID=' + restaurantID+ '&daysBack=' + daysBack, httpOptions );
+    }
+
     createArticle(token: any,userID:any,formValue:any): Observable<any> {
         const body = {
             name:formValue.name,
@@ -28,7 +38,7 @@ export class RestaurantService {
             price:formValue.price,
             restaurantID:userID,
             image:formValue.photo,
-            isDrink:false,
+            isDrink:formValue.isDrink,
         };
         const httpOptions = {
             headers: new HttpHeaders({
@@ -57,4 +67,43 @@ export class RestaurantService {
         };
         return this.http.post(`${environment.urlMenu}${API.createMenu}`, body, httpOptions );
     }
+    
+    getDrinks(token: any,idRestaurant:any): Observable<any> {
+        const httpOptions = {
+            headers: new HttpHeaders({
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`
+            })
+        };
+        return this.http.get(`${environment.urlProduct}${API.getDrinks}`+ '?restaurantID=' + idRestaurant, httpOptions );
+    }
+
+    getMenu(token: any,menuID:any): Observable<any> {
+        const body = {
+            id:menuID,
+        };
+        console.log(body);
+        const httpOptions = {
+            headers: new HttpHeaders({
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`
+            })
+        };
+        return this.http.post(`${environment.urlMenu}${API.getMenu}`, body, httpOptions );
+    }
+
+    getArticle(token: any,articleID:any): Observable<any> {
+        const body = {
+            id:articleID,
+        };
+        console.log(body);
+        const httpOptions = {
+            headers: new HttpHeaders({
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${token}`
+            })
+        };
+        return this.http.post(`${environment.urlProduct}${API.getArticle}`, body, httpOptions );
+    }
+
 }