+++ src/app/delivery/widget/homepage-delivery/homepage-delivery.component.ts
@@ -1,6 +1,8 @@
 import { HttpResponse } from '@angular/common/http';
 import { Component } from '@angular/core';
+import { ToastrService } from 'ngx-toastr';
 import { Delivery } from 'src/app/core/models/delivery.model';
+import { Orders } from 'src/app/core/models/infoOrders.model';
 import { DeliveryService } from 'src/app/core/services/delivery.service';
 import { SessionStorageService } from 'src/app/core/services/session-storage.service';
 
@@ -11,21 +13,31 @@ import { SessionStorageService } from 'src/app/core/services/session-storage.ser
 })
 export class HomepageDeliveryComponent {
   deliveryTest = new Delivery();
-  deliveries:Delivery[] = [];
+  deliveries: Orders[] = [];
 
-  constructor(private sessionStorageService: SessionStorageService,private deliveryService: DeliveryService) { }
+  constructor(private toastr: ToastrService, private sessionStorageService: SessionStorageService, private deliveryService: DeliveryService) { }
 
 
-  ngOnInit(){
+  ngOnInit() {
     let token = this.sessionStorageService.getItem('token');
-    this.deliveryService.getOrdersList(token).subscribe((response: HttpResponse<any>) => {
-      console.log(response)
+    this.deliveryService.getDeliveryWithFilter(token).subscribe({
+      next: (response: any) => {
+        const orders: Orders[] = response.orders.map((order: any) => ({
+          id: order._id,
+          clientID: order.clientID,
+          address: order.address,
+          date: order.date,
+          menus: order.menus,
+          status: order.status,
+          products: order.products,
+          totalPrice: order.totalPrice,
+          refusedBy: order.refusedBy
+        }));
+        this.deliveries = orders;    
+      },
+      error: () => {
+        this.toastr.error("Erreur lors de la récupération des commandes");
+      }
     });
-
-    this.deliveries.push(this.deliveryTest);
-    this.deliveryTest.desc="un composant super cool :]";
-    this.deliveryTest.name="Composant Cool";
-    this.deliveryTest.link="https://www.youtube.com/watch?v=dQw4w9WgXcQ";
-    this.deliveryTest.isAccepted=!true;
   }
 }