+++ src/app/restaurant/widget/add-menu/add-menu.component.ts
@@ -35,13 +35,15 @@ export class AddMenuComponent implements OnInit {
     drink: new FormControl(false),
     photo: new FormControl(''),
   });
-  
+
   ngOnInit(): void {
     this.userID = this.sessionStorageService.getItem('restaurantID');
     this.token = this.sessionStorageService.getItem('token');
     this.articlesInfoList = JSON.parse(this.sessionStorageService.getItem('articleList'));
     for (let i = 0; i < this.articlesInfoList.length; i++) {
-      this.articleList.push(this.articlesInfoList[i].name);
+      if (this.articlesInfoList[i].isDrink == false) {
+        this.articleList.push(this.articlesInfoList[i].name);
+      }
     }
   }
 
@@ -57,7 +59,7 @@ export class AddMenuComponent implements OnInit {
   }
 
   getIdFromArticle() {
-    let idList:string[] = [];
+    let idList: string[] = [];
     for (let i = 0; i < this.articleList.length; i++) {
       if (this.articleList.includes(this.menusForm.value.compo[i])) {
         const index = this.articleList.indexOf(this.menusForm.value.compo[i]);
@@ -71,16 +73,21 @@ export class AddMenuComponent implements OnInit {
 
   submitMenu() {
     let idList = this.getIdFromArticle();
-    this.restaurantService.createMenu(this.token, this.userID, this.menusForm.value,idList).subscribe((response: MessageModel) => {
-      if (response.message == "Menu added successfully") {
-        this.toastr.success("L'article a été ajouté avec succès");
-        setTimeout(() => {
-          this.router.navigate(['/restaurant']);
-        }, 2000);
+    this.restaurantService.createMenu(this.token, this.userID, this.menusForm.value, idList).subscribe({
+      next: (response: MessageModel) => {
+        if (response.message == "Menu added successfully") {
+          this.toastr.success("Le menu a été ajouté avec succès");
+          setTimeout(() => {
+            this.router.navigate(['/restaurant']);
+          }, 2000);
+        }
+        else {
+          this.toastr.error("Erreur lors de la création de l'article");
+        }
+      },
+      error: () => {
+        this.toastr.error("Erreur lors de la création du menu ");
       }
-      else {
-        this.toastr.error("Erreur lors de la création de l'article");
-      }
-    })
+    });
   }
 }