+++ src/app/sales/widget/user-card/user-card.component.ts
@@ -44,7 +44,8 @@ export class UserCardComponent {
     dialogRef.afterClosed().subscribe(result => {
       if (result === true) {
         console.log(this.token);
-        this.profilService.activateUser(this.token, this.menu.userID).subscribe((response: MessageModel) => {
+        this.profilService.activateUser(this.token, this.menu.userID).subscribe({
+      next: (response: MessageModel) => {
           if (response.message == 'User unsuspended') {
             this.toastr.success("L'utilisateur a été réactivé");
             setTimeout(() => {
@@ -54,7 +55,10 @@ export class UserCardComponent {
           else {
             this.toastr.error("Une erreur est survenue");
           }
-        });
+            },
+      error: () => {
+          this.toastr.error("Erreur lors de l'activation d'un utilisateur ");
+      }});
       }
     });
   }
@@ -68,8 +72,8 @@ export class UserCardComponent {
     });
     dialogRef.afterClosed().subscribe(result => {
       if (result === true) {
-        console.log(this.token);
-        this.profilService.disableUser(this.token, this.menu.userID).subscribe((response: MessageModel) => {
+        this.profilService.disableUser(this.token, this.menu.userID).subscribe({
+      next: (response: MessageModel) => {
           if (response.message == 'User suspended') {
             this.toastr.success("L'utilisateur a bien été suspendu");
             setTimeout(() => {
@@ -79,7 +83,10 @@ export class UserCardComponent {
           else {
             this.toastr.error("Une erreur est survenue");
           }
-        });
+            },
+      error: () => {
+          this.toastr.error("Erreur lors de la suspension de l'utilisateur ");
+      }});
       }
     });
   }
@@ -96,7 +103,8 @@ export class UserCardComponent {
     });
     dialogRef.afterClosed().subscribe(result => {
       if (result === true) {
-        this.profilService.deleteUser(this.token, this.menu.userID).subscribe((response: MessageModel) => {
+        this.profilService.deleteUser(this.token, this.menu.userID).subscribe({
+      next: (response: MessageModel) => {
           if (response.message == 'User deleted') {
             this.toastr.success("L'utilisateur a bien été supprimé");
             setTimeout(() => {
@@ -106,7 +114,10 @@ export class UserCardComponent {
           else {
             this.toastr.error("Une erreur est survenue");
           }
-        });
+            },
+      error: () => {
+          this.toastr.error("Erreur lors de la suppression de l'utilisateur ");
+      }});
       }
     });
   }