+++ src/app/technical/technical.component.ts
@@ -1,9 +1,11 @@
 import { Component } from '@angular/core';
-import { Router } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { LogService } from '../core/services/logs.service';
 import { SessionStorageService } from '../core/services/session-storage.service';
 import { HttpResponse } from '@angular/common/http';
 import { LogModel } from '../core/models/log.model';
+import { ComponentService } from '../core/services/component.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-technical',
@@ -11,26 +13,42 @@ import { LogModel } from '../core/models/log.model';
   styleUrls: ['./technical.component.css']
 })
 export class TechnicalComponent {
+  type!:string|null;
+  logs: string[][] = [];
+  token!: any;
+  log!: any;
+  constructor(private sessionStorageService: SessionStorageService,private route: ActivatedRoute,private toastr: ToastrService, private router: Router, private logservice: LogService, private sessionStorage: SessionStorageService, private composantService: ComponentService) { }
 
-  logs:string[]=[];
-  token!:any;
-  log!:any;
-  constructor(private router: Router,private logservice: LogService,private sessionStorage : SessionStorageService) { }
-
-  ngOnInit(){
+  ngOnInit() {
     this.token = this.sessionStorage.getItem("token");
 
     this.getLog();
+
+    this.type = this.sessionStorageService.getItem('type');
+    if(this.type != 'technical'){
+      this.router.navigate([`/error-page`], { relativeTo: this.route });
+    }
+    else{
+      this.type = "technical";
+    }
   }
 
-  getLog(){
-    this.log = this.logservice.getLog(this.token).subscribe((response: LogModel) => {
-      this.logs= response.logs;
+  getLog() {
+    this.log = this.logservice.getLog(this.token).subscribe({
+      next: (response: LogModel) => {
+        this.logs.push(response.logs);
+      },
+      error: () => {
+        this.toastr.error("Erreur lors de la récupération des logs ");
+      }
+    });
+    this.log = this.composantService.getLogComponent(this.token).subscribe((response: LogModel) => {
+      this.logs.push(response.logs);
     });
     //console.log(this.log);
   }
 
-  onClick(){
+  onClick() {
     console.log("test click");
     window.location.href = "http://localhost:8080";
   }