+++ src/environments/environment.ts
@@ -12,5 +12,7 @@ export const environment = {
     urlOrder: 'http://order.localhost',
     urlDelivery: 'http://delivery.localhost',
     urlMonitoring: 'http://monitoring.localhost',
+    urlComponent:'http://components.localhost',
+    
 
   };